+++ pdi/include/pdi/array_datatype.h
@@ -1,5 +1,6 @@
 /*******************************************************************************
  * Copyright (C) 2015-2019 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,7 +26,10 @@
 #ifndef PDI_ARRAY_DATATYPE_H_
 #define PDI_ARRAY_DATATYPE_H_
 
+#include <memory>
 #include <string>
+#include <utility>
+#include <vector>
 
 #include <pdi/pdi_fwd.h>
 #include <pdi/datatype.h>
@@ -52,6 +56,57 @@ class PDI_EXPORT Array_datatype:
 	size_t m_subsize;
 	
 public:
+	/** Accessor to get single element from array
+	 */
+	class Index_accessor : public Datatype::Accessor_base
+	{
+		/// Index of element that will be returned
+		size_t m_index;
+		
+		std::string access_kind() const override;
+	public:
+		/** Construct a new index accessor
+		 *
+		 * \param index index of element that will be returned
+		 */
+		Index_accessor(size_t index);
+		
+		std::pair<void*, Datatype_uptr> access(const Array_datatype& array_type,
+		    void* from,
+		    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_begin,
+		    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_end) const override;
+	};
+	
+	/** Accessor to get a slice of an array, returns array of the same subtype
+	 */
+	class Slice_accessor : public Datatype::Accessor_base
+	{
+		/// Start index of the slice
+		size_t m_start;
+		
+		/// End index of the slice
+		size_t m_end;
+		
+		/** Returns size of the slice
+		 *  \return size of the slice
+		 */
+		size_t size() const;
+		
+		std::string access_kind() const override;
+	public:
+		/** Construct a new slice accessor
+		 *
+		 * \param start start index of the slice
+		 * \param end end index of the slice
+		 */
+		Slice_accessor(size_t start, size_t end);
+		
+		std::pair<void*, Datatype_uptr> access(const Array_datatype& array_type,
+		    void* from,
+		    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_begin,
+		    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_end) const override;
+	};
+	
 	/** Construct a new partially filled Array_datatype
 	 *
 	 * \param[in] subtype the type of the elements contained in the array
@@ -113,6 +168,10 @@ public:
 	
 	void* data_from_dense_copy(void*, const void*) const override;
 	
+	std::pair<void*, Datatype_uptr> subaccess_by_iterators(void* from,
+	    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_begin,
+	    std::vector<std::unique_ptr<Accessor_base>>::const_iterator remaining_end) const override;
+	    
 	void destroy_data(void*) const override;
 	
 	std::string debug_string() const override;