+++ pdi/src/pointer_datatype.cxx
@@ -28,17 +28,46 @@
 #include <string>
 #include <sstream>
 
+#include "pdi/error.h"
+
 #include "pdi/pointer_datatype.h"
 
 
 namespace PDI {
 
+using std::endl;
+using std::function;
+using std::move;
+using std::pair;
+using std::string;
+using std::stringstream;
+using std::unique_ptr;
+using std::vector;
+
+string Pointer_datatype::Accessor::access_kind() const 
+{
+	return "pointer access";
+}
+
+pair<void*, Datatype_uptr> Pointer_datatype::Accessor::access(const Pointer_datatype& pointer_type,
+									void* from,
+									vector<unique_ptr<Accessor_base>>::const_iterator remaining_begin,
+									vector<unique_ptr<Accessor_base>>::const_iterator remaining_end) const
+{
+	from = *reinterpret_cast<void**>(from);
+	if (remaining_begin == remaining_end) {
+		return {from, pointer_type.subtype().clone_type()};
+	} else {
+		return pointer_type.subtype().subaccess_by_iterators(from, remaining_begin, remaining_end);
+	}
+}
+
 Pointer_datatype::Pointer_datatype(Datatype_uptr subtype):
-	m_subtype{std::move(subtype)}
+	m_subtype{move(subtype)}
 {}
 
-Pointer_datatype::Pointer_datatype(Datatype_uptr subtype, std::function<void* (void*, const void*)> copy, std::function<void(void*)> destroy):
-	m_subtype{std::move(subtype)},
+Pointer_datatype::Pointer_datatype(Datatype_uptr subtype, function<void* (void*, const void*)> copy, function<void(void*)> destroy):
+	m_subtype{move(subtype)},
 	m_copy{move(copy)},
 	m_destroy{move(destroy)}
 {}
@@ -55,12 +84,12 @@ Datatype_template_uptr Pointer_datatype::clone() const
 
 Datatype_uptr Pointer_datatype::clone_type() const
 {
-	return std::unique_ptr<Pointer_datatype> {new Pointer_datatype{m_subtype->clone_type(), m_copy, m_destroy}};
+	return unique_ptr<Pointer_datatype> {new Pointer_datatype{m_subtype->clone_type(), m_copy, m_destroy}};
 }
 
 Datatype_uptr Pointer_datatype::densify() const
 {
-	return std::unique_ptr<Pointer_datatype> {new Pointer_datatype{m_subtype->densify(), m_copy, m_destroy}};
+	return unique_ptr<Pointer_datatype> {new Pointer_datatype{m_subtype->densify(), m_copy, m_destroy}};
 }
 
 Datatype_uptr Pointer_datatype::evaluate(Context& ctx) const
@@ -109,6 +138,19 @@ void* Pointer_datatype::data_from_dense_copy(void* to, const void* from) const
 	return data_to_dense_copy(to, from);
 }
 
+
+pair<void*, Datatype_uptr> Pointer_datatype::subaccess_by_iterators(void* from,
+															vector<unique_ptr<Accessor_base>>::const_iterator remaining_begin,
+															vector<unique_ptr<Accessor_base>>::const_iterator remaining_end) const
+{
+	if (const Pointer_datatype::Accessor* pointer_accessor = dynamic_cast<const Pointer_datatype::Accessor*>(remaining_begin->get())) {
+		return remaining_begin->get()->access(*this, from, ++remaining_begin, remaining_end);
+	} else {
+		from = *reinterpret_cast<void**>(from);
+		return subtype().subaccess_by_iterators(from, remaining_begin, remaining_end);
+	}	
+}
+
 void Pointer_datatype::destroy_data(void* ptr) const
 {
 	if ( m_destroy ) {
@@ -116,15 +158,15 @@ void Pointer_datatype::destroy_data(void* ptr) const
 	}
 }
 
-std::string Pointer_datatype::debug_string() const
+string Pointer_datatype::debug_string() const
 {
-	std::stringstream ss;
-	ss << "type: pointer" << std::endl
-	    << "dense: " << (dense() ? "true" : "false") << std::endl
-	    << "buffersize: " << buffersize() << std::endl
-	    << "datasize: " << datasize() << std::endl
-	    << "alignment: " << alignment() << std::endl
-	    << "subtype: " << std::endl << m_subtype->debug_string();
+	stringstream ss;
+	ss << "type: pointer" << endl
+	    << "dense: " << (dense() ? "true" : "false") << endl
+	    << "buffersize: " << buffersize() << endl
+	    << "datasize: " << datasize() << endl
+	    << "alignment: " << alignment() << endl
+	    << "subtype: " << endl << m_subtype->debug_string();
 	return ss.str();
 }